corrected courses list pattern
