+++ application/smart_y/lib/bloc/courses_page/courses_page.bloc.dart
@@ -3,35 +3,39 @@ import 'package:inject/inject.dart';
 import 'package:rxdart/subjects.dart';
 import 'package:smarty/model/course/course_list_item.model.dart';
 import 'package:smarty/service/courses_page/courses_page.service.dart';
+import 'package:smarty/utils/logger/logger.dart';
 
 //State Management for courses page
-class CoursesPageBloc{
+@provide
+class CoursesPageBloc {
   static const int STATUS_CODE_INIT = -1;
   static const int STATUS_CODE_FETCHING_DATA = 566;
   static const int STATUS_CODE_FETCHING_DATA_ERROR = 458;
   static const int STATUS_CODE_FETCHING_DATA_SUCCESS = 758;
 
-  CoursesService _coursesService;
-  List<CourseListItem> courses;
-  CoursesPageBloc(this._coursesService);
+  final String tag = 'CoursesPageBloc';
 
-  PublishSubject<Pair<int, dynamic>> _coursesSubject = new PublishSubject();
-  Stream<Pair<int, dynamic>> get loginStateObservable => _coursesSubject.stream;
+  final CoursesService _coursesService;
+  final Logger _logger;
 
-  getCourses(){
+  CoursesPageBloc(this._coursesService, this._logger);
 
+  PublishSubject<Pair<int, List<CourseModel>>> _coursesSubject =
+      new PublishSubject();
+
+  Stream<Pair<int, List<CourseModel>>> get loginStateObservable =>
+      _coursesSubject.stream;
+
+  getCourses() {
     _coursesSubject.add(Pair(STATUS_CODE_FETCHING_DATA, null));
     _coursesService.getCourses().then((result) {
-      if (result != null ) {
-        courses = result;
-        _coursesSubject.add(Pair(STATUS_CODE_FETCHING_DATA_SUCCESS, null));
+      if (result != null) {
+        _coursesSubject.add(Pair(STATUS_CODE_FETCHING_DATA_SUCCESS, result));
+        _logger.info(tag, 'Data Fetched Correctly');
       } else {
-        print('fetching courses failed');
-        _coursesSubject.add(Pair(STATUS_CODE_FETCHING_DATA_ERROR,
-            "Error Fetching data, Please Try again Later"));
+        _coursesSubject.add(Pair(STATUS_CODE_FETCHING_DATA_ERROR, null));
+        _logger.error(tag, "Error Getting the Data");
       }
     });
   }
-
-
 }
