+++ application/smart_y/lib/bloc/courses_page/courses_page.bloc.inject.summary
@@ -1,17 +1,17 @@
 {
-  "asset": "asset:smarty/lib/main.dart",
+  "asset": "asset:smarty/lib/bloc/courses_page/courses_page.bloc.dart",
   "summary": {
     "injector": [],
     "module": [],
     "injectable": [
       {
-        "name": "MyApp",
+        "name": "CoursesPageBloc",
         "constructor": {
           "name": "",
           "kind": "constructor",
           "injectedType": {
             "lookupKey": {
-              "root": "asset:smarty/lib/main.dart#MyApp",
+              "root": "asset:smarty/lib/bloc/courses_page/courses_page.bloc.dart#CoursesPageBloc",
               "qualifier": null
             },
             "isProvider": false
@@ -21,14 +21,14 @@
           "dependencies": [
             {
               "lookupKey": {
-                "root": "asset:smarty/lib/ui/screen/login_page/login_page.dart#LoginPage",
+                "root": "asset:smarty/lib/service/courses_page/courses_page.service.dart#CoursesService",
                 "qualifier": null
               },
               "isProvider": false
             },
             {
               "lookupKey": {
-                "root": "asset:smarty/lib/ui/screen/courses_page/courses_page.dart#CoursesPage",
+                "root": "asset:smarty/lib/utils/logger/logger.dart#Logger",
                 "qualifier": null
               },
               "isProvider": false
