+++ application/smart_y/lib/manager/courses/cources.manager.dart
@@ -2,14 +2,15 @@ import 'package:analyzer_plugin/protocol/protocol.dart';
 import 'package:inject/inject.dart';
 import 'package:smarty/model/course/course_list_item.model.dart';
 import 'package:smarty/repository/courses_page/courses_page.repository.dart';
+import 'package:smarty/response/course_response/course_response.dart';
 
 @provide
-class CoursesManager{
+class CoursesManager {
   CoursesRepository _coursesRepository;
+
   CoursesManager(this._coursesRepository);
 
-  Future<List<CourseListItem>> getCourses()async{
-    var response = await _coursesRepository.getCourses();
-     return response;
+  Future<List<CourseResponse>> getCourses() async {
+    return this._coursesRepository.getCourses();
   }
 }
