+++ application/smart_y/lib/model/lesson/lesson.inject.summary
@@ -1,0 +1,8 @@
+{
+  "asset": "asset:smarty/lib/model/lesson/lesson.dart",
+  "summary": {
+    "injector": [],
+    "module": [],
+    "injectable": []
+  }
+}
