+++ application/smart_y/lib/response/course_response/course_response.inject.summary
@@ -1,0 +1,8 @@
+{
+  "asset": "asset:smarty/lib/response/course_response/course_response.dart",
+  "summary": {
+    "injector": [],
+    "module": [],
+    "injectable": []
+  }
+}
