+++ application/smart_y/lib/routes.dart
@@ -1,4 +1,5 @@
 class Routes {
   // This is a public class, every page has access to it in order to navigate to another screen
   static const LoginPageRoute = '/loginPage';
+  static const coursesPage = '/courses';
 }
