+++ application/smart_y/lib/service/courses_page/courses_page.service.dart
@@ -4,14 +4,24 @@ import 'package:smarty/model/course/course_list_item.model.dart';
 import 'package:smarty/persistence/shared_preferences/shared+preferences_helper.dart';
 import 'dart:convert';
 
+import 'package:smarty/response/course_response/course_response.dart';
+
 @provide
-class CoursesService{
+class CoursesService {
   SharedPreferencesHelper _sharedPreferencesHelper;
   CoursesManager _coursesManager;
-  CoursesService(this._sharedPreferencesHelper,this._coursesManager);
 
-  Future<List<CourseListItem>> getCourses()async{
+  CoursesService(this._sharedPreferencesHelper, this._coursesManager);
 
-      return _coursesManager.getCourses();
+  // Mapping data, and transforming it to useful data
+  Future<List<CourseModel>> getCourses() async {
+    List<CourseResponse> coursesResponse =
+        await this._coursesManager.getCourses();
+    List<CourseModel> availableCourses = [];
+    coursesResponse.forEach((course) {
+      // TODO: Create a Course Model
+      availableCourses.add(CourseModel(title: course.title.rendered, content: course.content.rendered));
+    });
+    return availableCourses;
   }
 }
