+++ application/smart_y/lib/ui/screen/course_details_page/Course_details_page.inject.summary
@@ -1,0 +1,8 @@
+{
+  "asset": "asset:smarty/lib/ui/screen/course_details_page/Course_details_page.dart",
+  "summary": {
+    "injector": [],
+    "module": [],
+    "injectable": []
+  }
+}
