+++ application/smart_y/lib/ui/screen/register_page/signup3_page.inject.summary
@@ -1,0 +1,8 @@
+{
+  "asset": "asset:smarty/lib/ui/screen/register_page/signup3_page.dart",
+  "summary": {
+    "injector": [],
+    "module": [],
+    "injectable": []
+  }
+}
