+++ application/smart_y/lib/ui/widget/app_drawer.inject.summary
@@ -1,0 +1,8 @@
+{
+  "asset": "asset:smarty/lib/ui/widget/app_drawer.dart",
+  "summary": {
+    "injector": [],
+    "module": [],
+    "injectable": []
+  }
+}
