+++ application/smart_y/lib/ui/widget/image_icon.inject.summary
@@ -1,0 +1,8 @@
+{
+  "asset": "asset:smarty/lib/ui/widget/image_icon.dart",
+  "summary": {
+    "injector": [],
+    "module": [],
+    "injectable": []
+  }
+}
