+++ application/smart_y/lib/ui/widget/offer_card.inject.summary
@@ -1,0 +1,8 @@
+{
+  "asset": "asset:smarty/lib/ui/widget/offer_card.dart",
+  "summary": {
+    "injector": [],
+    "module": [],
+    "injectable": []
+  }
+}
