+++ application/smart_y/lib/ui/widget/smarty_app_bar.inject.summary
@@ -1,0 +1,8 @@
+{
+  "asset": "asset:smarty/lib/ui/widget/smarty_app_bar.dart",
+  "summary": {
+    "injector": [],
+    "module": [],
+    "injectable": []
+  }
+}
