+++ application/smart_y/lib/utils/logger/logger.dart
@@ -1,3 +1,4 @@
+import 'package:fluttertoast/fluttertoast.dart';
 import 'package:inject/inject.dart';
 
 @provide
